bind UI for HeaterDetailFragment.kt.kt and update data

+++ app/src/main/java/com/example/mysmarthome/data/local/roomdatabase/DeviceEntity.kt
@@ -10,7 +10,7 @@ class DeviceEntity(
     val deviceName: String,
     var mode: DeviceMode?,
     var intensity: Int?,
-    var temperature: Int?,
+    var temperature: Float?,
     var position: Int?,
     var productType: ProductType
 ) {
